chore: move provisioner keys commands into slim build
